Moved pizza list to own component
