+++ src/components/PizzaList/PizzaList.jsx
@@ -1,0 +1,25 @@
+import axios from "axios";
+import { useEffect } from "react";
+
+export default function PizzaList() {
+
+  const fetchPizzas = () => {
+
+    axios.get('/api/pizza')
+      .then(res => {
+        console.log('got pizzas');
+
+      }).catch(err => console.log('fetchPizzas', err))
+
+  }
+
+  useEffect(() => {
+    fetchPizzas();
+  }, [])
+
+  return (
+    <div>
+
+    </div>
+  )
+}
